MXNet Extensions enhancements2 (#19016)
* initial commit

* split lib_api.h into lib_api.cc, updated examples for 2.0/gluon

* fixed licenses

* whitespace

* whitespace

* modernize

* fix modernize

* fix modernize

* fix modernize

* fixed move

* added lib_api.cc to CMakeLists.txt for example libs

* working example

* remove GLOBAL to fix protobuf issue

* fixed library unload

* added test target

* fixed sanity

* changed destructor to default

* added /LD option for customop_gpu_lib target

* moved /LD inside the <>

* diff compile flags for relu_lib.cu and lib_api.cc

* set CMAKE_VERBOSE_MAKEFILE for debug

* added -v to ninja

* added /MT

* another try

* changed /MT to -MT

* set flags for cxx separately

* split /LD /MT flags

* refactored cuda APIs into header file

* removed debugging stuff

* updated instructions for gpu build

* moved building into cmakelists

* moved build stuff into separate CMakeLists.txt

* fixed gpu example

* fixed license

* added dlmc library dependency

* added nnvm dependency

* removed nnvm dmlc dependencies, added WINDOWS_EXPORT_ALL_SYMBOLS option

* fixed WINDOWS_EXPORT_ALL_SYMBOLS

* changed nnvm to shared library

* backed out external ops changes

* split relu example into separate files to test separate lib_api.h/cc

* sanity

* addressed initial review items
