+++ config/linux_gpu.cmake
@@ -24,7 +24,8 @@
 #
 #  $ cp config/linux_gpu.cmake config.cmake
 #
-#  Next modify the according entries, and then compile by
+#  Next modify the entries in the config.cmake like MXNET_CUDA_ARCH to set the specific
+#  GPU architecture, and then compile by
 #
 #  $ mkdir build; cd build
 #  $ cmake ..
@@ -42,15 +43,18 @@ set(USE_CUDA ON CACHE BOOL "Build with CUDA support")
 set(USE_CUDNN ON CACHE BOOL "Build with cudnn support, if found")
 
 # Target NVIDIA GPU achitecture.
-# Valid options are "Auto" for autodetection, "All" for all available
-# architectures or a list of architectures by compute capability number, such as
-# "7.0" or "7.0;7.5" as well as name, such as "Volta" or "Volta;Turing".
+# Valid options are:
+#   - "Auto" for autodetection, will try and discover which GPU architecture to use by
+#            looking at the available GPUs on the machine that you're building on
+#   - "All" for all available GPU architectures supported by the version of CUDA installed
+#   - "specific GPU architectures" by giving the compute capability number such as
+#            "7.0" or "7.0;7.5" (ie. sm_70 or sm_75) or you can specify the name like:
+#            "Volta" or "Volta;Turing", be sure not to use quotes (ie. just set to 7.0)
 # The value specified here is passed to cmake's CUDA_SELECT_NVCC_ARCH_FLAGS to
 # obtain the compilation flags for nvcc.
 #
 # When compiling on a machine without GPU, autodetection will fail and you
-# should instead specify the target architecture manually to avoid excessive
-# compilation times.
+# should instead specify the target architecture manually.
 set(MXNET_CUDA_ARCH "Auto" CACHE STRING "Target NVIDIA GPU achitecture")
 
 #---------------------------------------------
