+++ example/extensions/lib_api/init_lib.cc
@@ -24,7 +24,7 @@
  */
 
 #include <iostream>
-#include "lib_api.h"
+#include "mxnet/lib_api.h"
 
 using namespace mxnet::ext;
