+++ example/extensions/lib_custom_op/test_transposecsr.py
@@ -54,25 +54,25 @@ print("data:", c.data.asnumpy())
 print("indices:", c.indices.asnumpy())
 print("indptr:", c.indptr.asnumpy())
 
-print("--------start symbolic compute--------")
+print("--------start Gluon compute--------")
 d = mx.sym.Variable('d')
 e = mx.sym.my_transposecsr(d)
 f = mx.sym.my_state_transposecsr(d, test_kw=200)
 
-exe = e.bind(ctx=mx.cpu(),args={'d':a})
-exe2 = f.bind(ctx=mx.cpu(),args={'d':a})
-out = exe.forward()
+block = mx.gluon.nn.SymbolBlock(e, [d])
+out = block(a)
 print("Compute Results:")
-print("data:", out[0].data.asnumpy())
-print("indices:", out[0].indices.asnumpy())
-print("indptr:", out[0].indptr.asnumpy())
+print("data:", out.data.asnumpy())
+print("indices:", out.indices.asnumpy())
+print("indptr:", out.indptr.asnumpy())
 
-out2 = exe2.forward()
-out2 = exe2.forward()
+block2 = mx.gluon.nn.SymbolBlock(f,[d])
+out2 = block2(a)
+out2 = block2(a)
 print("Stateful Compute Result:")
-print("data:", out2[0].data.asnumpy())
-print("indices:", out2[0].indices.asnumpy())
-print("indptr:", out2[0].indptr.asnumpy())
+print("data:", out2.data.asnumpy())
+print("indices:", out2.indices.asnumpy())
+print("indptr:", out2.indptr.asnumpy())
 
 print("--------Baseline(dense)--------")
 print(mx.nd.transpose(a.tostype('default')))
