+++ example/extensions/lib_custom_op/transposerowsp_lib.cc
@@ -25,7 +25,7 @@
 
 #include <iostream>
 #include <utility>
-#include "lib_api.h"
+#include "mxnet/lib_api.h"
 
 using namespace mxnet::ext;
