+++ src/initialize.cc
@@ -97,9 +97,7 @@ LibraryInitializer::LibraryInitializer()
   install_pthread_atfork_handlers();
 }
 
-LibraryInitializer::~LibraryInitializer() {
-  close_open_libs();
-}
+LibraryInitializer::~LibraryInitializer() = default;
 
 bool LibraryInitializer::lib_is_loaded(const std::string& path) const {
   return loaded_libs.count(path) > 0;
@@ -125,7 +123,13 @@ void* LibraryInitializer::lib_load(const char* path) {
       return nullptr;
     }
 #else
-    handle = dlopen(path, RTLD_LAZY);
+    /* library loading flags:
+     *  RTLD_LAZY - Perform lazy binding. Only resolve symbols as the code that
+     *              references them is executed.
+     *  RTLD_LOCAL - Symbols defined in this library are not made available to
+     *              resolve references in subsequently loaded libraries.
+     */
+    handle = dlopen(path, RTLD_LAZY | RTLD_LOCAL);
     if (!handle) {
       LOG(FATAL) << "Error loading library: '" << path << "'\n" << dlerror();
       return nullptr;
